Refactor Security Contexts:  - Move them to security_contexts package  - Switch to custom priorities for filter ordering management.
