+++ src/main/java/com/github/polimi_mt_acg/back2school/api/v1/administrators/AdministratorResource.java
@@ -1,5 +1,6 @@
 package com.github.polimi_mt_acg.back2school.api.v1.administrators;
 
+import com.github.polimi_mt_acg.back2school.api.v1.security_contexts.AdministratorSecured;
 import com.github.polimi_mt_acg.back2school.model.User;
 import com.github.polimi_mt_acg.back2school.model.User_;
 import com.github.polimi_mt_acg.back2school.utils.DatabaseHandler;
